Fix OpenCL compiler warnings for Intel GPU compatibility (issue #398)
- Cast printf argument to (unsigned int) in def_printFromWord macro to fix
  format specifier mismatch when word type is unsigned long (SHA-512 path)
- Applied fix in buffer_structs_template.cl, pbkdf2_sha1_32.cl, and
  pbkdf2_sha256_32.cl
- Add Intel Arc discrete GPU recognition in auto_select_opencl_platform

Agent-Logs-Url: https://github.com/3rdIteration/btcrecover/sessions/9213d4d6-5f73-4d89-9b8c-3b85c8aea5f0

+++ lib/opencl_brute/worker/generic/pbkdf2_sha1_32.cl
@@ -110,7 +110,7 @@ typedef struct {
             /* Prints little endian, since that's what we use */   \
             v = (v >> r) & 0xFF;                \
             if (hex) {                          \
-                printf("%02x", v);              \
+                printf("%02x", (unsigned int)v); \
             } else {                            \
                 printf("%c", (char)v);          \
             }                                   \
